Allow setting a maximum buffer ratio.
The alsa codec sets the maximum hardware buffer size to a large multiple of the
internal buffer size (32 by default) in order to minimize any chance of
underruns. In some cases, it may be useful to set a smaller ratio (say, 2) in
order to minimize latency. This commit allows globally setting the ratio using
the '-R' flag.