Allow loading effects files with '@' directive; update README.md.
It is now possible to construct the effects chain using files. The syntax is
@[~/]path. The '~/' prefix will be expanded to the contents of $HOME. On the
command line, the path is relative to $PWD. Within an effects file, the path
is relative to the directory containing said effects file. If the path starts
with a '/', the path is absolute.

The effects file syntax is very simple.
  * Arguments are delimited by whitespace.
  * Comments are specified using a '#' character at the beginning of a
    line (excluding any preceding spaces or tabs).
  * The '\' character removes any special meaning of the character that
    follows it.