Add 'sequence' mode.
In sequence mode, the inputs do not need to have the same sample rate or
number of channels. The effects chain and/or output will be rebuilt/reopened
when required.

This could use some (read: lots of) cleanup or a rewrite, as there is quite a
lot of redundant code.