resample.c: Do reflection/folding of the frequency spectrum.
I wasn't sure how to do this properly before, so I left it out. Without
the refection/folding, the error caused by the finite stopband attenuation
is sprayed throughout the frequency spectrum (but mostly concentrated around
nyquist). This doesn't matter (in terms of audibility) due to the extreme
stopband attenuation (~220dB) of the default filter, but doing the
reflection/folding is mathematically correct and allows the use of filters
with less stopband attenuation.