matrix4{,_mb}: Use laterally symmetric directional boost (yet again).
Independent boost distorts phantom source location somewhat when
panned between left/right and center, but eliminates pumping of the
opposite channel with hard-panned elements. What I'm not sure of is
whether this pumping is ever audible with normal program material
(i.e. not contrived test signals), and, if it is audible, which
defect is actually preferable.